SIG: Move sig.json parsing into a shared library
