+++ go/lib/sigjson/testdata/01-loadfromfile.json
